重够代码逻辑

+++ src/main/java/xyz/guqing/authorization/mapper/UserMapper.java
@@ -2,15 +2,15 @@ package xyz.guqing.authorization.mapper;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
-import xyz.guqing.authorization.entity.User;
-import xyz.guqing.authorization.entity.UserExample;
+import xyz.guqing.authorization.entity.model.User;
+import xyz.guqing.authorization.entity.model.UserExample;
 
 public interface UserMapper {
     int countByExample(UserExample example);
 
     int deleteByExample(UserExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(User record);
 
@@ -18,7 +18,7 @@ public interface UserMapper {
 
     List<User> selectByExample(UserExample example);
 
-    User selectByPrimaryKey(Integer id);
+    User selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") User record, @Param("example") UserExample example);
